% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickDataAPI.R
\name{bar}
\alias{bar}
\title{Fetch Bar Data}
\usage{
bar(dateRange, symbol, period)
}
\arguments{
\item{dateRange}{A character vector of length two \code{c("YYYY-MM-DD","YYYY-MM-DD")}.}

\item{symbol}{A character vector of length 1 that denotes a symbol.}

\item{period}{A character value is used with bar datatype to analyze data in different time segments.One of "\code{1min}","\code{5min}","\code{1hour}" and "\code{1day}".}
}
\value{
A data frame object is returned that contains open,high,low,close,quantity and weighted average price columns.
}
\description{
Fetches  bar data from historical data service
}
\seealso{
\code{\link{trade}}, \code{\link{bestbidoffer}}, \code{\link{depth}}, \code{\link{openinterest}}.
}

