% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickDataAPI.R
\name{bestbidoffer}
\alias{bestbidoffer}
\title{Fetch Bestbidoffer Data}
\usage{
bestbidoffer(dateRange, symbol)
}
\arguments{
\item{dateRange}{A character vector of length two \code{c("YYYY-MM-DD","YYYY-MM-DD")}.}

\item{symbol}{A character vector of length 1 that denotes a symbol.}
}
\value{
A data frame object is returned that contains best bid and offer prices with quantity.
}
\description{
Fetches bestbidoffer data from historical data service
}
\seealso{
\code{\link{trade}}, \code{\link{depth}}, \code{\link{openinterest}}, \code{\link{bar}}.
}

