% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickDataAPI.R
\name{depth}
\alias{depth}
\title{Fetchn Depth Data}
\usage{
depth(dateRange, symbol)
}
\arguments{
\item{dateRange}{A character vector of length two \code{c("YYYY-MM-DD","YYYY-MM-DD")}.}

\item{symbol}{A character vector of length 1 that denotes a symbol.}
}
\value{
A data frame object is returned with row number,action,bid or ask,order count,price and quantity columns.
}
\description{
Fetches depth data from historical data service
}
\details{
Symbols in \bold{BISTPP} and \bold{BISTVIOP}, are available on depth dataType.
}
\seealso{
\code{\link{trade}}, \code{\link{bestbidoffer}}, \code{\link{openinterest}}, \code{\link{bar}}.
}

