% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickDataAPI.R
\name{getData}
\alias{getData}
\title{Fetch Data}
\usage{
getData(dateRange, symbol, type, period = NULL)
}
\arguments{
\item{dateRange}{A character vector of length two \code{c("YYYY-MM-DD","YYYY-MM-DD")}.}

\item{symbol}{A character vector of length 1 that denotes a symbol.}

\item{period}{A character value is used with bar datatype to analyze data in different time segments.One of "\code{1min}","\code{5min}","\code{1hour}" and "\code{1day}".}

\item{dataType}{A character value to analyze various types of data .One of " \code{trade}"," \code{bestbidoffer}"," \code{depth}", " \code{openinterest}" and " \code{bar}".}
}
\value{
A data frame object is returned in "\code{trade}","\code{bestbidoffer}","\code{depth}","\code{openinterest}" and "\code{bar}" dataTypes.
}
\description{
Fetches data from historical data service
}
\seealso{
\code{\link{trade}}, \code{\link{bestbidoffer}}, \code{\link{depth}}, \code{\link{openinterest}}, \code{\link{bar}}
}

