% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriks.R
\name{matriksData}
\alias{matriksData}
\title{matriksData}
\usage{
matriksData(ticker, dataType, startDate, endDate = NULL, period = NULL)
}
\arguments{
\item{ticker}{A character variable for ticker symbol.}

\item{dataType}{A character type variable. One of the followings: "\code{trade}","\code{bestbidoffer}","\code{depth}", "\code{openinterest}", or "\code{bar}".}

\item{startDate}{A character object as "\code{YYYY-MM-DD}"}

\item{endDate}{A character object as "\code{YYYY-MM-DD}"}

\item{period}{A character variable. Valid only when datatype arguement is set to "\code{bar}", otherwise ignored. One of the following character variables: "\code{1min}","\code{5min}","\code{1hour}" , or "\code{1day}".}
}
\value{
A data.frame object is is returned. See details above.
}
\description{
Fetches various price data from Matriks Historical Data Service
}
\details{
A \code{matriksData} object is a data frame that contains \code{trade}, \code{bestbidoffer}, \code{depth}, \code{openinterest}
 and \code{bar}.

If dataType is \code{trade}, data frame contains seven column: \code{timestamp},
\code{symbol}, \code{bid_or_ask}, \code{trade_no}, \code{price}, \code{quantity} and \code{signal_time}. \code{timestamp} column includes
transaction time.\code{bid_or_ask} column delineates whether the trade occured on the b(id) or on the a(sk) side. \code{signal_time} shows the number of seconds elapsed
since \code{01/01/1970}.

By setting dataType to \code{bestbidoffer}  one has the following columns:
six columns:\code{timestamp},\code{symbol},\code{best_bid_price},\code{best_bid_size},\code{best_ask_price},\code{best_ask_size}.

If dataType is set to \code{openinterest}, total amount of nominal positions on options or futures are returned.

If dataType is set to \code{bar}, the function fetches aggregated prices (open, high, close, low) and total volume information for the chosen \code{period} level of aggregation.
}
\note{
Although, as of now, Matriks Data Services can provide with historical data since the year 2013, the data coverage is more likely to expand in time.
}
\examples{
##NOT RUN
matriksData("GARAN","bar","2016-01-22","2016-01-23","1hour")
matriksData("GARAN","trade","2016-01-22","2016-01-23")
##NOT RUN
}
\seealso{
\code{\link{trade}}, \code{\link{bestbidoffer}}, \code{\link{depth}}, \code{\link{openinterest}}, \code{\link{bar}}
}

