% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataAPI.R
\name{matriksExchanges}
\alias{matriksExchanges}
\title{Fetch Exchanges}
\usage{
matriksExchanges()
}
\value{
A data.frame is returned that contains  exchanges' code and description  on Matriks.
}
\description{
Fetches exchange data from historical data service
}
\details{
All of exchanges on Matriks, is returned.There is not any classification.
}
\seealso{
\code{\link{matriksSymbols}},\code{\link{matriksMarkets}}, \code{\link{matriksMarketsViop}}
}

