% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataAPI.R
\name{matriksMarkets}
\alias{matriksMarkets}
\title{Fetch Markets}
\usage{
matriksMarkets()
}
\value{
A data.frame is returned that contains markets' code and descrioption on Matriks.
}
\description{
Fetches market data from historical data service
}
\details{
All of markets on Matriks, is returned.There is not any classification.
}
\seealso{
\code{\link{matriksSymbols}},\code{\link{matriksExchanges}}, \code{\link{matriksMarketsViop}}
}

