% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataAPI.R
\name{matriksMarketsViop}
\alias{matriksMarketsViop}
\title{Fetch Futures and Options Markets}
\usage{
matriksMarketsViop()
}
\value{
A data.frame is returned that contains futures and options markets on Matriks with market code, description and id.
}
\description{
Fetches futures and options market data from historical data service
}
\details{
All of futures and options markets on Matriks, is returned.There is not any classification.
}
\seealso{
\code{\link{matriksSymbols}},\code{\link{matriksExchange}},\code{\link{matriksMarkets}}
}

