% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataAPI.R
\name{matriksSymbols}
\alias{matriksSymbols}
\title{Fetch Symbols}
\usage{
matriksSymbols(exchange_id = NULL, market = NULL, state = NULL)
}
\arguments{
\item{exchange_id}{A numeric value or NULL(meaning all) of exchanges.}

\item{market}{A character value or NULL(meaning all) of markets.}

\item{state}{A character value or NULL(meaning all) of  symbol activity.One of "active" ,"passive" and "all".

\deqn{You can enter the default arguments(NULL) 'exchange_id', 'market' and 'state' as empty arrays [ ].}}
}
\value{
A data.frame is returned that  contains symbols on Matriks.
}
\description{
Fetches symbol data from historical data service
}
\seealso{
\code{\link{matriksExchanges}},\code{\link{matriksMarkets}}, \code{\link{matriksMarketsViop}}
}

