% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickDataAPI.R
\name{openinterest}
\alias{openinterest}
\title{Fetch Openinterest Data}
\usage{
openinterest(dateRange, symbol)
}
\arguments{
\item{dateRange}{A character vector of length two \code{c("YYYY-MM-DD","YYYY-MM-DD")}.}

\item{symbol}{A character vector of length 1 that denotes a symbol.}
}
\value{
An xts object is returned that contains numbers of open positions.
}
\description{
Fetches openinterest data from historical data service
}
\details{
"Symbols in BISTVIOP","FUTURES" and "SERBEST", are available on openinterest dataType.
}
\seealso{
\code{\link{trade}}, \code{\link{bestbidoffer}}, \code{\link{depth}},\code{\link{bar}}.
}

