% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickDataAPI.R
\name{trade}
\alias{trade}
\title{Fetch Trade Data}
\usage{
trade(dateRange, symbol)
}
\arguments{
\item{dateRange}{A character vector of length two  \code{c("YYYY-MM-DD","YYYY-MM-DD")}.}

\item{symbol}{A character vector of length 1 that denotes a symbol.}
}
\value{
A data frame object is returned that contains bid or ask,trade id, price quantity and signal time columns.
}
\description{
Fetches trade data from historical data service.
}
\seealso{
\code{\link{bestbidoffer}}, \code{\link{depth}}, \code{\link{openinterest}}, \code{\link{bar}}.
}

